pyproject.toml
[project]
name = "worker-bot"
version = "0.1.0"
description = "Bot de Telegram que expone agentes Aider sobre varios proyectos y permite tareas proactivas via cron."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aider-chat>=0.86.2",
    "apscheduler>=3.11.2",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.7",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["workerbot"]
